pyproject.toml
[project]
name = "researchproposalagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.4.1",
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.21",
    "langchain-google-genai>=4.2.1",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.12",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.3",
    "openai>=2.29.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.42.0",
]
